#10074: Move digamma_bw to TTNN
