+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -836,8 +836,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.erf_bw
 
-.. autofunction:: tt_lib.tensor.digamma_bw
-
 .. autofunction:: tt_lib.tensor.deg2rad_bw
 
 .. autofunction:: tt_lib.tensor.threshold_bw
