+++ docs/source/ttnn/ttnn/ttnn/digamma_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.digamma_bw:
+
+ttnn.digamma_bw
+###############
+
+.. autofunction:: ttnn.digamma_bw
