+++ docs/source/ttnn/ttnn/ttnn/reciprocal_bw.rst
@@ -1,6 +1,6 @@
 .. _ttnn.reciprocal_bw:
 
 ttnn.reciprocal_bw
-###############
+##################
 
 .. autofunction:: ttnn.reciprocal_bw
