+++ tests/ttnn/unit_tests/operations/backward/test_backward_digamma.py
@@ -4,8 +4,8 @@
 
 import torch
 import pytest
-import tt_lib
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import compare_pcc, data_gen_with_range
+import ttnn
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import compare_pcc, data_gen_with_range
 
 
 @pytest.mark.parametrize(
@@ -22,7 +22,7 @@ def test_bw_digamma(input_shapes, device):
 
     pyt_y = torch.digamma(in_data)
 
-    tt_output_tensor_on_device = tt_lib.tensor.digamma_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.digamma_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
