+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -157,11 +157,6 @@ std::vector<Tensor> erf_bw(
     const Tensor& input,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> digamma_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
 std::vector<Tensor> deg2rad_bw(
     const Tensor& grad,
     const Tensor& input,
